Using floats for now.

+++ db/migrate/20150204103958_create_cities.rb
@@ -2,8 +2,8 @@ class CreateCities < ActiveRecord::Migration
   def change
     create_table :cities do |t|
       t.string :name
-      t.decimal :lat, {:precision=>10, :scale=>6}
-      t.decimal :lng, {:precision=>10, :scale=>6}
+      t.float :lat, {:precision=>10, :scale=>6}
+      t.float :lng, {:precision=>10, :scale=>6}
 
       t.timestamps
     end
